Create roles.php
Sorry lads in the UK. I'll be back on Thursday, hopefully this works for you guys. For future reference use BOOLEAN instead of bit because boolean data counts as a tinyint TRUE=1 FALSE=0. The bit doesn't function as a tinyint creating a lot of unnecessary busywork. Talk to you later.